+++ nens_auth_client/tests/test_middleware.py
@@ -1,6 +1,7 @@
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser
 from nens_auth_client.middleware import AccessTokenMiddleware
+
 import pytest