+++ nens_auth_client/users.py
@@ -8,6 +8,7 @@ from django.utils.crypto import get_random_string
 
 import logging
 
+
 logger = logging.getLogger(__name__)